move todo to models dir
